Support arrays for local variables
 * Don't yet support globals
 * Don't support function params because MSDN doesn't say they exist
   (although fxc accepts it)
 * Only support literal array dimensions

+++ src/clc/cir.rs
@@ -46,7 +46,7 @@ pub enum Type {
     UIntPtrT,
     Struct(Identifier),
     Pointer(AddressSpace, Box<Type>),
-    Array(Box<Type>, u32),
+    Array(Box<Type>, u64),
 
     Image1D(AccessModifier),
     Image1DBuffer(AccessModifier),
